Get current CITES appendix listings and reservations.

Retrieve current CITES appendix listings and reservations, CITES quotas, and CITES suspensions for a given taxon concept.

spp_cites_legislation(taxon_id, scope = "current", language = "en",
  raw = FALSE, token = NULL, verbose = TRUE, ...)

Arguments

taxon_id
a vector of character strings containing species' taxon concept identifiers (see spp_taxonconcept()).

scope
vector of character strings indicating the time scope of legislation, values are taken among current, historic and all. Default is current.

language
vector of character strings indicating the language for the text of legislation notes, values are taken among en (English), fr (French) and es (Spanish). Default is en.

raw
a logical. Should raw data be returned?

token
a character string containing the authentification token, see https://api.speciesplus.net/documentation. Default is set to NULL and requires the environment variable SPECIESPLUS_TOKEN to be set directly in Renviron. Alternatively, set_token() can be used to set SPECIESPLUS_TOKEN for the current session.

verbose
a logical. Should extra information be reported on progress?

...
Further named parameters, see httr::GET().

Value

If raw is set to TRUE then an object of class spp_raw (or spp_raw_multi if length(taxon_id)>1) is returned which is essentially a list of lists (see option as = 'parsed' in httr::content()). Otherwise, an object of class spp_cites_leg (or spp_cites_leg_multi if length(taxon_id)>1) is returned which is a list of three data frames:

cites_listings: lists CITES annex listings EU suspensions,

cites_quotas: lists CITES quotas,

cites_suspensions: lists CITES suspensions.

References

https://api.speciesplus.net/documentation/v1/cites_legislation/index.html

Examples

res1 <- spp_cites_legislation(taxon_id = 4521)

#> >>> Now processing taxon_id '4521'............

#> Error in rcites_res(q_url, token, ...): Unauthorized (HTTP 401).

res2 <- spp_cites_legislation(taxon_id = c('4521', '3210', '10255'))

#> >>> Now processing taxon_id '4521'............

#> Error in rcites_res(q_url, token, ...): Unauthorized (HTTP 401).

res3 <- spp_cites_legislation(taxon_id = 4521, scope = 'all',
verbose = FALSE, config=httr::verbose())

#> Error in rcites_res(q_url, token, ...): Unauthorized (HTTP 401).

res4 <- spp_cites_legislation(taxon_id = 4521, language = 'fr')

#> >>> Now processing taxon_id '4521'............

#> Error in rcites_res(q_url, token, ...): Unauthorized (HTTP 401).